buildsys: detet antlr3 availability
antlr3 was silently required and a build would fail without explanation
if it was missing. Add FindANTLR cmake module to look for the library
and give a proper warning if not found.

Additionally, introduce cmake modules sub-directory. Now that there are
already two modules it seems to make sense -- more will probably follow.