FawkesBackend: loggig improvements
Log less, explicit log calls now are prefixed with "[YAGI]". Explicitly
log which events occured. For now, do not log game_state events as they
are too frequent. This is an LLSF-specific thing which can be removed
once we can have the generic integration (YAGI requires non-finite
domains for that).