llsf2015: add interfaces for reading
The current agent has to make use of the following interfaces:
"RobotinoLightInterface" in order to receive the light signals gathered
by the responsible vision.
"ZoneInterface" in order to receive further information from executing
the explore_zone skill. This interface serves the information whether an
MPS is detected in the particular zone and (if there is an MPS detected)
the particular tag_id seen on the MPS.