FawkesBackend: add lightspecs_to_types map
Due to YAGI's limitations we cannot store arbitrary data in any way.
Although the LLSF's rules from 2015 require us to store an arbitrary
string according to a light signal. To handle this we make use of a map
which stores light-signals (e.g. in the form of "ONOFFON") according to strings.