FawkesBackend: pass lightspecs_to_types
The lightspecs_to_types-map is passed to the constructors of the
ExogenousEventProducer and the SignalHandler.
The ExogenousEventProducer has to fill the map with the according values
given by the protobuf messages.
The SignalHandler is then supposed to use this map in order to correctly
announce the <type> matching to the machine name. This is done in the
agent by sending the machine-name and its according light_states.